+++ src/Dap.Build/RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 * Save original version before inject
 * Create Recover target to remove the injected version
 * Create sub tragets per project
+* Add Run and WatchRun targets
+* Add Fable targets
 
 ## 0.6.2
 * Create targets for each projects